Fix db won't be updated after set/clear buyer.
If you set/clear winBidder then change to
another session and back to current session.
The bidder will be changed to the value before
modified. This is caused by we don't refresh
the m_collectionList in m_aeInternet.
Now if we changned session, we will clear
m_collectionList first to force getting the auctions'
data again.